feat: 시간별로 로그를 한 줄 요약하는 로직 구현

+++ src/main/java/com/example/llmn/controller/DTO/MetricResponse.java
@@ -20,7 +20,7 @@ public class MetricResponse {
 
     public record MemoryMetricDTO(
             String time,
-            long memoryUsage) {}
+            double memoryUsage) {}
 
     public record NetworkInMetricDTO(
             String time,
